Separate benchmark files in order to add more functionality (SRP)
